app/requirements.txt
aiohttp==3.8.4
apscheduler==3.10.1
PyYAML==6.0
SQLAlchemy==2.0.15
asyncpg==0.27.0
alembic==1.11.1
aiohttp-cors==0.7.0
aiohttp-apispec==2.2.3
marshmallow==3.19.0